+++ PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,17 @@
+# [{card_number}] {title}
+
+## 노션 칸반 링크
+
+- [카드 제목](url)
+
+## 카드에서 구현 혹은 해결하려는 내용
+
+-
+
+## 테스트 방법
+
+-
+
+## 기타 사항
+
+- 고민되는 부분 기록
