+++ src/constants/constants.js
@@ -1,9 +1,24 @@
-export const markdownExample = `# Styled-Markdown
-  ## 사용법 소개  
-  왼쪽에 마크다운 방식으로 문서를 작성해보세요.
-  오른쪽에 변형된 형태로 출력됩니다.
+export const markdownExample = `
+  # Styled-Markdown
+  Styled-Markdown에 오신 것을 환영합니다.
+  기본적인 사용법을 알려 드릴게요.
 
-  ### 강조하기
-  지정된 스타일로 강조하고 싶은 부분은 느낌표로 텍스트를 감싸보세요.
-  !!이렇게요!!!
+  ## 기본 사용법
+  왼쪽에 마크다운 문법으로 글을 쓰시면 오른쪽에 결과물이 나타납니다.
+  얼마든지 원하는 대로 쓰고 수정해 보세요.
+
+  ## 제목은 이렇게 쓰세요
+  - 제목은 #부터 ######까지 총 6가지를 지원합니다.
+  - 리스트는 이렇게 쓰시면 됩니다.
+    * 별표로 표시할 수도 있어요.
+    * 탭을 쓰면 이렇게 되죠.
+
+  그 외 기본적인 마크다운 문법을 지원합니다.
+  *이탤릭*은 별표 하나로, **볼드체**는 별표 두 개로 감싸면 됩니다.
+  ~~취소선~~은 물결 두 개로 감싸면 되죠.
+
+  ## 중요한 문자는 이렇게 써 보세요.
+  !!강조해야 하는 문자열!!은 느낌표 두 개로 감싸보세요. 기본 붉은 색과 볼드체 스타일을 더합니다.
+  색상을 변경하고 싶다면 %blue_이렇게 색상 키워드와 함께% 텍스트를 써 보세요.
+  배경색으로 하이라이트를 더하고 싶다면 %%yellow_이렇게 쓸 수도%% 있답니다.
 `;
